+++ app/containers/wallet/WalletSettingsPage.js
@@ -4,6 +4,7 @@ import { observer, inject } from 'mobx-react';
 import WalletSettings from '../../components/wallet/WalletSettings';
 import type { InjectedProps } from '../../types/injectedPropsType';
 import type { AssuranceMode } from '../../types/transactionAssuranceTypes';
+import { isValidWalletName } from '../../lib/validations';
 
 @inject('stores', 'actions') @observer
 export default class WalletSettingsPage extends Component {
@@ -19,6 +20,17 @@ export default class WalletSettingsPage extends Component {
     const { wallets, walletSettings, uiDialogs } = this.props.stores;
     const { actions } = this.props;
     const activeWallet = wallets.active;
+    const {
+      updateWalletRequest,
+      lastUpdatedWalletField,
+      walletFieldBeingEdited,
+    } = walletSettings;
+    const {
+      startEditingWalletField,
+      stopEditingWalletField,
+      cancelEditingWalletField,
+      updateWalletField,
+    } = actions.walletSettings;
 
     // Guard against potential null values
     if (!activeWallet) throw new Error('Active wallet required for WalletSettingsPage.');
@@ -33,6 +45,16 @@ export default class WalletSettingsPage extends Component {
         isWalletPasswordSet={activeWallet.hasPassword}
         walletPasswordUpdateDate={activeWallet.passwordUpdateDate}
         isDialogOpen={uiDialogs.isOpen}
+        walletName={activeWallet.name}
+        isSubmitting={updateWalletRequest.isExecuting}
+        isInvalid={updateWalletRequest.wasExecuted && updateWalletRequest.result === false}
+        lastUpdatedField={lastUpdatedWalletField}
+        onFieldValueChange={(field, value) => updateWalletField.trigger({ field, value })}
+        onStartEditing={field => startEditingWalletField.trigger({ field })}
+        onStopEditing={stopEditingWalletField.trigger}
+        onCancelEditing={cancelEditingWalletField.trigger}
+        activeField={walletFieldBeingEdited}
+        nameValidator={name => isValidWalletName(name)}
       />
     );
   }